Fix bug#  1620646 - wbemsmt-webapp: indicator for longrunning actions / confirmation for critical actions

+++ java/org/sblim/wbemsmt/cli/dns/DeleteDnsAddressMatchList.java
@@ -145,8 +145,16 @@ public class DeleteDnsAddressMatchList extends CimCommand {
 			
 			dc.trace(System.out);
 			
-			adapter.delete();
-			System.out.println("\n" + bundle.getString("deleted",new Object[]{bundle.getString("DnsAddressMatchListForServiceDataContainer.caption")}));
+			if (getConfirmation())
+			{
+    			adapter.delete();
+    			System.out.println("\n" + bundle.getString("deleted",new Object[]{bundle.getString("DnsAddressMatchListForServiceDataContainer.caption")}));
+			}
+			else
+			{
+    			System.out.println("\n" + bundle.getString("not.deleted",new Object[]{bundle.getString("DnsAddressMatchListForServiceDataContainer.caption")}));
+			}
+			
 		
 		}
 		catch (Exception e)
