+++ java/org/sblim/wbemsmt/cli/dns/DeleteDnsResourceRecord.java
@@ -165,8 +165,16 @@ public class DeleteDnsResourceRecord extends CimCommand {
 			
 			dc.trace(System.out);
 			
-			adapter.delete();
-			System.out.println("\n" + bundle.getString("deleted",new Object[]{bundle.getString("DnsResourceRecordDataContainer.caption")}));
+			if (getConfirmation())
+			{
+    			adapter.delete();
+    			System.out.println("\n" + bundle.getString("deleted",new Object[]{bundle.getString("DnsResourceRecordDataContainer.caption")}));
+			}
+			else
+			{
+    			System.out.println("\n" + bundle.getString("not.deleted",new Object[]{bundle.getString("DnsResourceRecordDataContainer.caption")}));
+			}
+			
 		
 		}
 		catch (Exception e)
